db/pyproject.toml
[build-system]
build-backend = "setuptools.build_meta"
requires = ["setuptools>=61"]

[project]
authors = [{email = "dcs@psych.gatech.edu", name = "dcs"}]
dependencies = [
  "sqlalchemy>=2.0.0",
  "asyncpg>=0.29.0",
  "aiosqlite>=0.20.0",
  "alembic>=1.13.0",
  "python-dotenv>=1.0.0",
  "greenlet>=3.3.2",
]
description = "Database layer for the DCS simulation engine."
license = "MIT"
name = "dcs-db"
requires-python = ">=3.11"
version = "0.1.0"

[project.optional-dependencies]
dev = [
  "pytest>=8.0.0",
  "pytest-asyncio>=0.23.0",
]

[tool.setuptools.packages.find]
include = ["dcs_db*"]
